fix (carousel): resize detection using polling
Resize detection using the "on resize" event has a few issues

* If the window is resized progressively, many resizes occur one after
  the other, causing a great CPU usage
* For some reason, if you resize your window too much, it ends up
  stuck at a given size, without being correctly updated
* If the resize is due to another DOM element changing size, without
  the window being resize, it will not be detected

The safest way here to detect a resize is to actually poll for size
changes. If the size is stable for 200ms, then the resize procedure is
done.

This has a very, very low CPU overhead (it's one number comparison every
100ms, which is not really a challenge for the JS interpreter)